# Steamfitter.Api.Client/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Steamfitter API",
    "version": "v1"
  },
  "paths": {
...
    "/api/scenarios": {
      "get": {
        "tags": [
          "Scenario"
        ],
        "summary": "Gets all Scenario in the system",
        "description": "Returns a list of all of the Scenarios in the system.\r\n<para />\r\nOnly accessible to a SuperUser",
        "operationId": "getScenarios",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Scenario"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Scenario"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Scenario"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Scenario"
        ],
        "summary": "Creates a new Scenario",
        "description": "Creates a new Scenario with the attributes specified\r\n<para />\r\nAccessible only to a SuperUser or an Administrator",
        "operationId": "createScenario",
        "requestBody": {
          "description": "The data to create the Scenario with",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Scenario"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Scenario"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Scenario"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Scenario"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Scenario"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Scenario"
                }
              }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/scenarios/{id}": {
      "get": {
        "tags": [
          "Scenario"
        ],
        "summary": "Gets a specific Scenario by id",
        "description": "Returns the Scenario with the id specified\r\n<para />\r\nAccessible to a SuperUser or a User that is a member of a Team within the specified Scenario",
        "operationId": "getScenario",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id of the Scenario",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The id of the Scenario",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Scenario"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Scenario"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Scenario"
                }
              }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Scenario"
        ],
        "summary": "Updates a Scenario",
        "description": "Updates an Scenario with the attributes specified\r\n<para />\r\nAccessible only to a SuperUser or a User on an Admin Team within the specified Scenario",
        "operationId": "updateScenario",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The Id of the Exericse to update",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The Id of the Exericse to update",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "The updated Scenario values",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Scenario"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Scenario"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Scenario"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Scenario"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Scenario"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Scenario"
                }
              }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Scenario"
        ],
        "summary": "Deletes an Scenario",
        "description": "Deletes an Scenario with the specified id\r\n<para />\r\nAccessible only to a SuperUser or a User on an Admin Team within the specified Scenario",
        "operationId": "deleteScenario",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id of the Scenario to delete",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The id of the Scenario to delete",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/scenarios/me": {
      "get": {
        "tags": [
          "Scenario"
        ],
        "summary": "Gets the personal Scenario for the current user",
        "description": "Returns the current user's personal Scenario\r\n<para />\r\nAccessible to an authenticated User",
        "operationId": "getMyScenario",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Scenario"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Scenario"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Scenario"
                }
              }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/scenariotemplates/{id}/scenarios": {
      "post": {
        "tags": [
          "Scenario"
        ],
        "summary": "Creates a new Scenario from a ScenarioTemplate",
        "description": "Creates a new Scenario from the specified ScenarioTemplate\r\n<para />\r\nAccessible only to a SuperUser or an Administrator",
        "operationId": "createScenarioFromScenarioTemplate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The ScenarioTemplate ID to create the Scenario with",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The ScenarioTemplate ID to create the Scenario with",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Scenario"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Scenario"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Scenario"
                }
              }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/scenarios/{id}/copy": {
      "post": {
        "tags": [
          "Scenario"
        ],
        "summary": "Creates a new Scenario from a Scenario",
        "description": "Creates a new Scenario from the specified Scenario\r\n<para />\r\nAccessible only to a SuperUser or an Administrator",
        "operationId": "copyScenario",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The Scenario ID to copy into a new Scenario",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The Scenario ID to copy into a new Scenario",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Scenario"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Scenario"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Scenario"
                }
              }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/scenarios/{id}/users/add": {
      "put": {
        "tags": [
          "Scenario"
        ],
        "summary": "Adds Users to a Scenario",
        "description": "Adds one or more Users to the Scenario specified\r\n<para />\r\nAccessible only to a SuperUser ContentDeveloper",
        "operationId": "addUsersToScenario",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "description": "",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "",
                "nullable": true
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "",
                "nullable": true
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "",
                "nullable": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Scenario"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Scenario"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Scenario"
                }
              }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/scenarios/{id}/start": {
      "put": {
        "tags": [
          "Scenario"
        ],
        "summary": "Start a Scenario",
        "description": "Updates a Scenario to active and executes initial Tasks\r\n<para />\r\nAccessible only to a SuperUser or a User on an Admin Team within the specified Scenario",
        "operationId": "startScenario",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The Id of the Scenario to update",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The Id of the Scenario to update",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Scenario"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Scenario"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Scenario"
                }
              }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/scenarios/{id}/pause": {
      "put": {
        "tags": [
          "Scenario"
        ],
        "summary": "Pause a Scenario",
        "description": "Updates a Scenario to paused\r\n<para />\r\nAccessible only to a SuperUser or a User on an Admin Team within the specified Scenario",
        "operationId": "pauseScenario",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The Id of the Scenario to update",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The Id of the Scenario to update",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Scenario"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Scenario"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Scenario"
                }
              }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ScenarioStatus": {
        "enum": [
          "ready",
          "active",
          "paused",
          "ended",
          "archived",
          "error"
        ],
        "type": "string"
      },
...
      "Scenario": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "$ref": "#/components/schemas/ScenarioStatus"
          },
          "onDemand": {
            "type": "boolean"
          },
          "scenarioTemplateId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "viewId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "view": {
            "type": "string",
            "nullable": true
          },
          "defaultVmCredentialId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "vmCredentials": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VmCredential"
            },
            "nullable": true
          },
          "users": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          },
          "dateCreated": {
            "type": "string",
            "format": "date-time"
          },
          "dateModified": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdBy": {
            "type": "string",
            "format": "uuid"
          },
          "modifiedBy": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "implicit": {
            "authorizationUrl": "http://localhost:5000/connect/authorize",
            "scopes": {
              "steamfitter player player-vm": "public api access"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "oauth2": [
        "steamfitter player player-vm"
      ]
    }
  ]
}